collectors: cut the process about time to tsutil module
