+++ collectors/metrics/bin/get_metrics_from_prom.py
@@ -35,16 +35,12 @@
 """
 
 import argparse
-import concurrent.futures
-import datetime
 import json
 import sys
-import urllib.error
-import urllib.parse
-import urllib.request
 from typing import Any
 
 from prometheus.fetcher import PromFetcher
+from tsutil import tsutil
 
 COMPONENT_LABELS = {
     "front-end", "orders", "orders-db", "carts", "carts-db",
@@ -55,38 +51,7 @@ APP_LABEL = 'sock-shop'
 APP_NODEPOOL = 'default-pool'
 STEP = 15
 DEFAULT_DURATION = "30m"
-NAN = 'nan'
 GRAFANA_DASHBOARD = "d/3cHU4RSMk/sock-shop-performance"
-
-
-def interpotate_time_series(
-    values: list[list[int]], time_meta: dict[str, Any],
-) -> list[list[float]]:
-    start, end, step = time_meta['start'], time_meta['end'], time_meta['step']
-    new_values = []
-
-    # start check
-    if (lost_num := int((values[0][0] - start) / step) - 1) > 0:
-        for j in range(lost_num):
-            new_values.append([start + step * j, NAN])
-
-    for i, val in enumerate(values):
-        if i + 1 >= len(values):
-            new_values.append(val)
-            break
-        cur_ts, next_ts = val[0], values[i + 1][0]
-        new_values.append(val)
-        if (lost_num := int((next_ts - cur_ts) / step) - 1) > 0:
-            for j in range(lost_num):
-                new_values.append([cur_ts + step * (j + 1), NAN])
-
-    # end check
-    last_ts = values[-1][0]
-    if (lost_num := int((end - last_ts) / step)) > 0:
-        for j in range(lost_num):
-            new_values.append([last_ts + step * (j + 1), NAN])
-
-    return new_values
 
 
 def support_set_default(obj: set):
@@ -137,7 +102,7 @@ def metrics_as_result(
         data['containers'].setdefault(container, [])
         metric_name = labels['__name__']
 
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'container_name': container,
             'metric_name': metric_name,
@@ -171,7 +136,7 @@ def metrics_as_result(
         # TODO: want to rename 'kubernetes-name' in prometheus config because it is not intuitive.
         container = metric['metric']['kubernetes_name']
         data['middlewares'].setdefault(container, [])
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'container_name': container,
             'metric_name': metric['metric']['__name__'],
@@ -185,7 +150,7 @@ def metrics_as_result(
             continue
         node = metric['metric']['node']
         data['nodes'].setdefault(node, [])
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'node_name': node,
             'metric_name': metric['metric']['__name__'],
@@ -196,7 +161,7 @@ def metrics_as_result(
     for metric in throughput_metrics:
         service = metric['metric']['name']
         data['services'].setdefault(service, [])
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'service_name': metric['metric']['name'],
             'metric_name': 'throughput',
@@ -207,7 +172,7 @@ def metrics_as_result(
     for metric in error_metrics:
         service = metric['metric']['name']
         data['services'].setdefault(service, [])
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'service_name': metric['metric']['name'],
             'metric_name': 'errors',
@@ -218,7 +183,7 @@ def metrics_as_result(
     for metric in latency_metrics:
         service = metric['metric']['name']
         data['services'].setdefault(service, [])
-        values = interpotate_time_series(metric['values'], time_meta)
+        values = tsutil.interpotate_time_series(metric['values'], time_meta)
         m = {
             'service_name': metric['metric']['name'],
             'metric_name': 'latency',
@@ -246,53 +211,6 @@ def metrics_as_result(
     return data
 
 
-def get_unix_time(timestamp: Any) -> int:
-    if timestamp.isdigit():  # check unix time
-        return int(timestamp)
-    if timestamp is None or not timestamp:
-        return 0
-    dt = datetime.datetime.strptime(timestamp, '%Y-%m-%dT%H:%M:%SZ')
-    return int(dt.timestamp())
-
-
-def time_range_from_args(params: dict[str, Any]) -> tuple[int, int]:
-    """
-    get unix timestamps range (start to end) from duration
-    """
-
-    duration = datetime.timedelta(seconds=0)
-    dt = params["duration"]
-    if dt.endswith("s") or dt.endswith("sec"):
-        duration = datetime.timedelta(seconds=int(dt[:-1]))
-    elif dt.endswith("m") or dt.endswith("min"):
-        duration = datetime.timedelta(minutes=int(dt[:-1]))
-    elif dt.endswith("h") or dt.endswith("hours"):
-        duration = datetime.timedelta(hours=int(dt[:-1]))
-    else:
-        raise ValueError("args.duration is invalid format")
-
-    duration = int(duration.total_seconds())
-    now = int(datetime.datetime.now().timestamp())
-    start, end = now - duration, now
-    if params["end"] is None and params["start"] is None:
-        pass
-    elif params["end"] is not None and params["start"] is None:
-        end = get_unix_time(params["end"])
-        start = end - duration
-    elif params["end"] is None and params["start"] is not None:
-        start = get_unix_time(params["start"])
-        end = start + duration
-    elif params["end"] is not None and params["start"] is not None:
-        start = get_unix_time(params["start"])
-        end = get_unix_time(params["end"])
-    else:
-        raise ValueError("not reachable")
-
-    start = start - start % params["step"]
-    end = end - end % params["step"]
-    return start, end
-
-
 def collect_metrics(
     prometheus_url: str, grafana_url: str, start_time: str, end_time: str, chaos_param: dict[str, str],
     duration: str = DEFAULT_DURATION, step: int = STEP,
@@ -301,7 +219,7 @@ def collect_metrics(
     Collect metrics API
     """
 
-    start, end = time_range_from_args({
+    start, end = tsutil.time_range_from_args({
         "duration": duration,
         "start": start_time,
         "end": end_time,
@@ -440,6 +358,9 @@ def collect_metrics(
 
 def main():
     parser = argparse.ArgumentParser()
+    parser.add_argument("--target",
+                        choices=['sockshop', 'trainticket'], default='sockshop',
+                        help="target application to collect metrics")
     parser.add_argument("--prometheus-url",
                         help="endpoint URL for prometheus server",
                         default="http://localhost:9090")
